pyproject.toml
[project]
name = "paperscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-functions>=1.22.1",
    "azure-storage-blob>=12.25.1",
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "logging>=0.4.9.6",
    "openai>=1.71.0",
    "pymupdf>=1.25.5",
    "requests>=2.32.3",
    "typing>=3.10.0.0",
    "uvicorn>=0.34.0",
]
